+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * change: Sequencex.flattenSequenceOfIterable() rename to flattenOfIterable
 * fix: Fix the stack overflow bug when Sequencex.flatten() is passed into TransformingSequence
 * new: Added 'tools4j-resources' module and 'tools4j-run' module
+* new: Sequencex add flatMapOfIterable, flatMapOfIterableTo, flatMapIndexed, flatMapIndexedOfIterable, flatMapIndexedOfIterableTo, flatMapIndexedTo method
 
 ## v1.0.0-beta02
 improve: Improve code